Fix #1789 Add action to return back to the All resources page from a viewer (#1791)
