+++ geonode_mapstore_client/client/js/routes/Viewer.jsx
@@ -113,32 +113,6 @@ function ViewerRoute({
     const Loader = loaderComponent;
     const className = `page-${resourceType}-viewer`;
 
-    useEffect(() => {
-        // set the correct height of navbar
-        const mainHeader = document.querySelector('.gn-main-header');
-        const mainHeaderPlaceholder = document.querySelector('.gn-main-header-placeholder');
-        const topbar = document.querySelector('#gn-topbar');
-        function resize() {
-            if (mainHeaderPlaceholder && mainHeader) {
-                mainHeaderPlaceholder.style.height = mainHeader.clientHeight + 'px';
-            }
-            if (topbar && mainHeader) {
-                topbar.style.top = mainHeader.clientHeight + 'px';
-            }
-        }
-        // hide the navigation bar if a resource is being viewed
-        if (!loading) {
-            document.getElementById('gn-topbar')?.classList.add('hide-navigation');
-            document.getElementById('gn-brand-navbar-bottom')?.classList.add('hide-search-bar');
-            resize();
-        }
-        return () => {
-            document.getElementById('gn-topbar')?.classList.remove('hide-navigation');
-            document.getElementById('gn-brand-navbar-bottom')?.classList.remove('hide-search-bar');
-            resize();
-        };
-    }, [loading]);
-
     return (
         <>
             {resource && <MetaTags
