+++ geonode_mapstore_client/static/mapstore/configs/localConfig.json
@@ -530,6 +530,12 @@
                     "containerPosition": "header",
                     "leftMenuItems": [
                         {
+                            "type": "link",
+                            "size": "md",
+                            "href": "#/all",
+                            "labelId": "gnviewer.allResources"
+                        },
+                        {
                             "type": "plugin",
                             "size": "md",
                             "name": "DetailViewerButton"
@@ -1476,6 +1482,12 @@
                     "containerPosition": "header",
                     "leftMenuItems": [
                         {
+                            "type": "link",
+                            "size": "md",
+                            "href": "#/all",
+                            "labelId": "gnviewer.allResources"
+                        },
+                        {
                             "type": "plugin",
                             "size": "md",
                             "name": "DetailViewerButton"
@@ -2206,6 +2218,12 @@
                     "containerPosition": "header",
                     "leftMenuItems": [
                         {
+                            "type": "link",
+                            "size": "md",
+                            "href": "#/all",
+                            "labelId": "gnviewer.allResources"
+                        },
+                        {
                             "type": "plugin",
                             "size": "md",
                             "name": "DetailViewerButton"
@@ -2528,6 +2546,12 @@
                     "containerPosition": "header",
                     "leftMenuItems": [
                         {
+                            "type": "link",
+                            "size": "md",
+                            "href": "#/all",
+                            "labelId": "gnviewer.allResources"
+                        },
+                        {
                             "type": "plugin",
                             "size": "md",
                             "name": "DetailViewerButton"
@@ -2778,6 +2802,12 @@
                 "cfg": {
                     "containerPosition": "header",
                     "leftMenuItems": [
+                        {
+                            "type": "link",
+                            "size": "md",
+                            "href": "#/all",
+                            "labelId": "gnviewer.allResources"
+                        },
                         {
                             "type": "plugin",
                             "size": "md",
@@ -4083,6 +4113,13 @@
                 "cfg": {
                     "containerPosition": "header",
                     "leftMenuItems": [
+                        {
+                            "type": "link",
+                            "size": "md",
+                            "href": "#/all",
+                            "labelId": "gnviewer.allResources",
+                            "disableIf": "{!!state('resourceParams')?.mapPk}"
+                        },
                         {
                             "type": "link",
                             "href": "{'#/map/' + state('resourceParams')?.mapPk}",
