+++ geonode_mapstore_client/static/mapstore/gn-translations/data.de-DE.json
@@ -404,7 +404,8 @@
                         "linkResource": "Die verknüpfte Ressource konnte nicht gespeichert werden"
                     }
                 }
-            }
+            },
+            "allResources": "All resources"
         }
     }
 }
