+++ geonode_mapstore_client/static/mapstore/gn-translations/data.en-US.json
@@ -404,7 +404,8 @@
                         "linkResource": "Failed to save linked resource"
                     }
                 }
-            }
+            },
+            "allResources": "All resources"
         }
     }
 }
