+++ geonode_mapstore_client/static/mapstore/gn-translations/data.es-ES.json
@@ -403,7 +403,8 @@
                         "linkResource": "No se pudo guardar el recurso vinculado"
                     }
                 }
-            }
+            },
+            "allResources": "All resources"
         }
     }
 }
