+++ geonode_mapstore_client/static/mapstore/gn-translations/data.fr-FR.json
@@ -404,7 +404,8 @@
                         "linkResource": "Échec de l'enregistrement de la ressource liée"
                     }
                 }
-            }
+            },
+            "allResources": "All resources"
         }
     }
 }
