+++ geonode_mapstore_client/static/mapstore/gn-translations/data.it-IT.json
@@ -406,7 +406,8 @@
                         "linkResource": "Impossibile salvare la risorsa collegata"
                     }
                 }
-            }
+            },
+            "allResources": "All resources"
         }
     }
 }
