+++ geonode_mapstore_client/static/mapstore/gn-translations/data.sv-SE.json
@@ -374,7 +374,8 @@
                         "linkResource": "Failed to save linked resource"
                     }
                 }
-            }
+            },
+            "allResources": "All resources"
         }
     }
 }
