+++ geonode_mapstore_client/static/mapstore/gn-translations/data.zh-ZH.json
@@ -373,7 +373,8 @@
                         "linkResource": "Failed to save linked resource"
                     }
                 }
-            }
+            },
+            "allResources": "All resources"
         }
     }
 }
